move code into segment mesh controller and ensure selected super voxel stays highlighted after proofread operation
