+++ frontend/javascripts/oxalis/view/plane_view.ts
@@ -38,60 +38,6 @@ const raycaster = new VisibilityAwareRaycaster();
 let oldRaycasterHit: MeshSceneNode | null = null;
 const MESH_HOVER_THROTTLING_DELAY = 150;
 
-const ACTIVATED_COLOR = [0.7, 0.5, 0.1] as const;
-const HOVERED_COLOR = [0.65, 0.5, 0.1] as const;
-
-function updateMeshAppearance(
-  _mesh: THREE.Object3D,
-  isHovered: boolean | undefined,
-  isActiveUnmappedSegment?: boolean | undefined,
-) {
-  const mesh = _mesh as MeshSceneNode;
-  let wasChanged = false;
-  if (isHovered != null) {
-    if (!!mesh.isHovered !== isHovered) {
-      mesh.isHovered = isHovered;
-      wasChanged = true;
-    }
-  }
-  if (isActiveUnmappedSegment != null) {
-    if (!!mesh.isActiveUnmappedSegment !== isActiveUnmappedSegment) {
-      mesh.isActiveUnmappedSegment = isActiveUnmappedSegment;
-      wasChanged = true;
-    }
-  }
-  if (!wasChanged) {
-    // Nothing to do
-    return;
-  }
-
-  const targetOpacity = mesh.isHovered ? 0.8 : 1.0;
-  mesh.parent.traverse((child) => {
-    if (child instanceof THREE.Mesh) {
-      child.material.opacity = targetOpacity;
-    }
-  });
-  if (mesh.isHovered || mesh.isActiveUnmappedSegment) {
-    mesh.material.emissive.setHSL(...HOVERED_COLOR);
-
-    if (mesh.material.savedHex == null) {
-      mesh.material.savedHex = mesh.material.color.getHex();
-    }
-    const newColor: readonly [number, number, number] = mesh.isHovered
-      ? HOVERED_COLOR
-      : ACTIVATED_COLOR;
-    mesh.material.color.setHSL(...newColor);
-    mesh.material.opacity = 1.0;
-  } else {
-    // @ts-ignore
-    mesh.material.emissive.setHex("#FF00FF");
-    if (mesh.material.savedHex != null) {
-      mesh.material.color.setHex(mesh.material.savedHex);
-    }
-    mesh.material.savedHex = undefined;
-  }
-}
-
 class PlaneView {
   cameras: OrthoViewMap<THREE.OrthographicCamera>;
   throttledPerformMeshHitTest: (
@@ -190,7 +136,8 @@ class PlaneView {
   ): RaycastIntersection<THREE.Object3D> | null | undefined {
     const storeState = Store.getState();
     const SceneController = getSceneController();
-    const { meshesLODRootGroup } = SceneController.segmentMeshController;
+    const { segmentMeshController } = SceneController;
+    const { meshesLODRootGroup } = segmentMeshController;
     const tdViewport = getInputCatcherRect(storeState, "TDView");
     const { hoveredSegmentId } = storeState.temporaryConfiguration;
 
@@ -226,7 +173,7 @@ class PlaneView {
 
     // Undo highlighting of old hit
     if (oldRaycasterHit?.parent != null) {
-      updateMeshAppearance(oldRaycasterHit, false);
+      segmentMeshController.updateMeshAppearance(oldRaycasterHit, false);
 
       oldRaycasterHit = null;
     }
@@ -235,7 +182,7 @@ class PlaneView {
 
     // Highlight new hit
     if (hitObject?.parent != null) {
-      updateMeshAppearance(hitObject, true);
+      segmentMeshController.updateMeshAppearance(hitObject, true);
 
       Store.dispatch(
         updateTemporarySettingAction(
@@ -285,6 +232,9 @@ class PlaneView {
   }
 
   start(): void {
+    const SceneController = getSceneController();
+    const { segmentMeshController } = SceneController;
+
     this.unsubscribeFunctions.push(
       app.vent.on("rerender", () => {
         this.needsRerender = true;
@@ -319,23 +269,11 @@ class PlaneView {
           }
           return segmentationTracing.activeUnmappedSegmentId;
         },
-        (activeUnmappedSegmentId) => this.highlightUnmappedSegmentId(activeUnmappedSegmentId),
+        (activeUnmappedSegmentId) =>
+          segmentMeshController.highlightUnmappedSegmentId(activeUnmappedSegmentId),
         true,
       ),
     );
-  }
-  highlightUnmappedSegmentId(activeUnmappedSegmentId: number | null | undefined): void {
-    const SceneController = getSceneController();
-    const { meshesLODRootGroup } = SceneController.segmentMeshController;
-    meshesLODRootGroup.traverse((_obj) => {
-      // The cast is save because MeshSceneNode adds only optional properties
-      const obj = _obj as MeshSceneNode;
-      if (activeUnmappedSegmentId != null && obj.unmappedSegmentId === activeUnmappedSegmentId) {
-        updateMeshAppearance(obj, undefined, true);
-      } else if (obj.isActiveUnmappedSegment) {
-        updateMeshAppearance(obj, undefined, false);
-      }
-    });
   }
 
   getCameraForPlane(plane: Viewport) {
